refactor!: migrate to new `@aws-sdk/client-s3`
