+++ index.js
@@ -1,15 +1,8 @@
-/*
-
-	node-utils by SWR audio lab
-	frequently used node-packages and helpers
-
-*/
-
 // import packages
 const ard = require('./packages/ard')
 const helpers = require('./packages/helpers')
 const numbers = require('./packages/numbers')
-const storage = require('./packages/storage-wrapper')
+const storage = require('./packages/storage')
 const strings = require('./packages/strings')
 const undici = require('./packages/undici')
