+++ packages/storage/index.js
@@ -1,0 +1,7 @@
+// import package
+const storage = require('../../utils/storage')
+
+// export package
+module.exports = {
+	storage,
+}
