+++ packages/undici/index.js
@@ -1,5 +1,13 @@
-// load utils
-const request = require('./request')
+/*
 
-// export handler with tracing, if enabled
-module.exports = (tracer) => tracer?.wrap('undici.request', request) || request
+	by SWR Audio Lab
+
+	this file runs the undici-wrapper util
+
+*/
+
+// load request handler
+const undici = require('../../utils/undici/index')
+
+// export handler
+module.exports = undici()
